backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
psycopg2==2.9.9
python-dotenv==1.0.0
groq==0.4.2
requests==2.31.0
aiohttp==3.9.1
